backend/apis/auth_service/requirements.txt
asgiref==3.8.1
Django==4.2.11
sqlparse==0.4.4
django-cors-headers==4.3.1
psycopg2==2.9.6
python-dotenv==0.10.1
MarkupSafe==2.1.5
PyJWT==2.8.0
Werkzeug==3.0.2
requests==2.31.0
requests-oauthlib==2.0.0
cryptography==41.0.0
pyotp==2.8.0